Fix: bug in -r -rp option - provided by Charles P. Wright from IBM Research
git-svn-id: svn://svn.code.sf.net/p/sipp/code/sipp/trunk@133 60913f6b-29f6-41b8-8391-1ad2d2d6c344

+++ sipp.cpp
@@ -3532,7 +3532,7 @@ int main(int argc, char *argv[])
 	}
 	switch(option->type) {
 	  case SIPP_OPTION_FLOAT_TO_SEC:
-	    *((int *)option->data) /= 1000;
+	    *((double *)option->data) /= 1000;
 	    break;
 	}
 	break;
